requirements.txt
matplotlib
numpy
pandas
Pillow
scikit_learn
torch
torchvision
timm
termcolor
yacs